requirements.txt
requests>=2.31.0
flask>=3.0.0
paho-mqtt>=2.0.0
pandas>=2.0.0
matplotlib>=3.7.0
numpy>=1.24.0
seaborn>=0.12.0

